+++ Tests/ServerTests/ThrowableRouteTests.swift
@@ -26,7 +26,7 @@ class ThrowableRouteTests: TestCase {
     }
 
     func setup(
-        port: UInt16,
+        port: Int,
         serverCode: @escaping (Server) throws -> Void,
         clientCode: @escaping (Client) throws -> Void
     ) {
